extract term.EditString
