+++ cmd/exo/state.go
@@ -2,16 +2,12 @@ package main
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io/ioutil"
 	"os"
-	"os/exec"
-	"strings"
 
 	"github.com/deref/exo/internal/core/api"
 	"github.com/deref/exo/internal/util/jsonutil"
-	"github.com/deref/exo/internal/util/which"
+	"github.com/deref/exo/internal/util/term"
 	"github.com/spf13/cobra"
 )
 
@@ -118,74 +114,8 @@ var stateEditCmd = &cobra.Command{
 			return err
 		}
 
-		tmpfile, err := ioutil.TempFile("", "state.*.json")
-		if err != nil {
-			return fmt.Errorf("creating temporary file: %w", err)
-		}
-		defer os.Remove(tmpfile.Name())
-
-		if err := jsonutil.PrettyPrintJSONString(tmpfile, output.State); err != nil {
-			return fmt.Errorf("writing to temporary file: %w", err)
-		}
-		if err := tmpfile.Close(); err != nil {
-			return fmt.Errorf("closing temporary file: %w", err)
-		}
-
-		stat, err := os.Stat(tmpfile.Name())
-		if err != nil {
-			return fmt.Errorf("checking modification time: %w", err)
-		}
-		originalModTime := stat.ModTime()
-
-		editor := os.Getenv("EDITOR")
-		if editor == "" {
-
-			for _, candidateEditor := range []string{
-				"sensible-editor",
-				"editor",
-				"code",
-				"vim",
-				"nano",
-				"vi",
-				"emacs",
-				"ee",
-			} {
-				found, err := which.Which(candidateEditor)
-				if err != nil && !strings.Contains(err.Error(), "not found") {
-					return fmt.Errorf("looking up candidate editor %q: %w", candidateEditor, err)
-				}
-				if found != "" {
-					editor = found
-					break
-				}
-			}
-
-			if editor == "" {
-				return errors.New("No editor available")
-			}
-		}
-
-		edit := exec.Command(editor, tmpfile.Name())
-		edit.Stdin = os.Stdin
-		edit.Stdout = os.Stdout
-		edit.Stderr = os.Stderr
-		if err := edit.Run(); err != nil {
-			return fmt.Errorf("editing state file: %w", err)
-		}
-
-		if stat, err = os.Stat(tmpfile.Name()); err != nil {
-			return fmt.Errorf("checking modification time: %w", err)
-		}
-		newModTime := stat.ModTime()
-		if !newModTime.After(originalModTime) {
-			fmt.Fprintf(os.Stderr, "Not modified - not updating state.")
-			return nil
-		}
-
-		newState, err := os.ReadFile(tmpfile.Name())
-		if err != nil {
-			return fmt.Errorf("reading updated state: %w", err)
-		}
+		oldState := output.State
+		newState, err := term.EditString("state.*.json", oldState)
 
 		_, err = workspace.SetComponentState(ctx, &api.SetComponentStateInput{
 			Ref:   componentRef,
